Refactored internal export and host function code
