+++ sdk/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/wasm/host/WasmHostFunctions.scala
@@ -1,0 +1,32 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.digitalasset.daml.lf.speedy.wasm
+
+package host
+
+import com.digitalasset.daml.lf.data.Ref
+import com.digitalasset.daml.lf.value.{Value => LfValue}
+
+import scala.concurrent.duration.Duration
+
+trait WasmHostFunctions {
+
+  def logInfo(msg: String): Unit
+
+  def createContract(templateCons: Ref.TypeConRef, args: LfValue): LfValue.ContractId
+
+  def fetchContractArg(
+      templateId: Ref.TypeConRef,
+      contractId: LfValue.ContractId,
+      timeout: Duration,
+  ): LfValue
+
+  def exerciseChoice(
+      templateId: Ref.TypeConRef,
+      contractId: LfValue.ContractId,
+      choiceName: Ref.ChoiceName,
+      choiceArg: LfValue,
+      consuming: Boolean,
+  ): LfValue
+}
